chore: refactor guest page
