+++ src/components/guest/GuestPagePagination.tsx
@@ -1,0 +1,130 @@
+import {
+  Pagination,
+  PaginationContent,
+  PaginationEllipsis,
+  PaginationItem,
+  PaginationLink,
+  PaginationNext,
+  PaginationPrevious,
+} from "@/components/ui/pagination";
+import { cn } from "@/lib/utils";
+
+interface GuestPagePaginationProps {
+  currentPage: number;
+  totalPages: number;
+  handlePageChange: (page: number) => void;
+}
+
+export default function GuestPagePagination({
+  handlePageChange,
+  currentPage,
+  totalPages,
+}: GuestPagePaginationProps) {
+  const renderPaginationItems = () => {
+    const items = [];
+
+    // Helper function to add page number
+    const addPageNumber = (pageNum: number) => {
+      items.push(
+        <PaginationItem key={pageNum}>
+          <PaginationLink
+            onClick={() => handlePageChange(pageNum)}
+            isActive={currentPage === pageNum}
+            className={cn(
+              "min-w-9 rounded-md",
+              currentPage === pageNum
+                ? "pointer-events-none bg-[#8B4513] text-white hover:bg-[#A0522D]"
+                : "text-[#8B4513] hover:bg-[#8B4513]/10",
+            )}
+          >
+            {pageNum}
+          </PaginationLink>
+        </PaginationItem>,
+      );
+    };
+
+    // Always show first page
+    addPageNumber(1);
+
+    if (totalPages <= 5) {
+      // If 5 or fewer pages, show all
+      for (let i = 2; i <= totalPages; i++) {
+        addPageNumber(i);
+      }
+    } else {
+      if (currentPage <= 3) {
+        // Near start
+        for (let i = 2; i <= 4; i++) {
+          addPageNumber(i);
+        }
+        items.push(
+          <PaginationItem key="end-ellipsis">
+            <PaginationEllipsis className="text-[#8B4513]" />
+          </PaginationItem>,
+        );
+      } else if (currentPage >= totalPages - 2) {
+        // Near end
+        items.push(
+          <PaginationItem key="start-ellipsis">
+            <PaginationEllipsis className="text-[#8B4513]" />
+          </PaginationItem>,
+        );
+        for (let i = totalPages - 3; i < totalPages; i++) {
+          addPageNumber(i);
+        }
+      } else {
+        // In middle
+        items.push(
+          <PaginationItem key="start-ellipsis">
+            <PaginationEllipsis className="text-[#8B4513]" />
+          </PaginationItem>,
+        );
+        for (let i = currentPage - 1; i <= currentPage + 1; i++) {
+          addPageNumber(i);
+        }
+        items.push(
+          <PaginationItem key="end-ellipsis">
+            <PaginationEllipsis className="text-[#8B4513]" />
+          </PaginationItem>,
+        );
+      }
+
+      // Always show last page
+      if (totalPages > 1) {
+        addPageNumber(totalPages);
+      }
+    }
+
+    return items;
+  };
+
+  return (
+    <div className="mt-6">
+      <Pagination>
+        <PaginationContent className="gap-2">
+          <PaginationItem>
+            <PaginationPrevious
+              onClick={() => handlePageChange(Math.max(1, currentPage - 1))}
+              className={cn(
+                "rounded-md border-[#8B4513] text-[#8B4513] hover:bg-[#8B4513]/10",
+                currentPage === 1 && "pointer-events-none opacity-50",
+              )}
+            />
+          </PaginationItem>
+          {renderPaginationItems()}
+          <PaginationItem>
+            <PaginationNext
+              onClick={() =>
+                handlePageChange(Math.min(totalPages, currentPage + 1))
+              }
+              className={cn(
+                "rounded-md border-[#8B4513] text-[#8B4513] hover:bg-[#8B4513]/10",
+                currentPage === totalPages && "pointer-events-none opacity-50",
+              )}
+            />
+          </PaginationItem>
+        </PaginationContent>
+      </Pagination>
+    </div>
+  );
+}
